Issue #1: Where are the classes EclipseRendering, PlanetaryRendering, SkyRendering ?
They must be defined in the package jparsec.graph.chartRendering.frame, but they are not in the branch master.

Compared with the content of http://conga.oan.es/\~alonso/jparsec/lib/jparsec.jar and extracted the required Java sources. removed the duplicate classes.

Resolved, by introducing the classes and eliminating the duplicate classes.